don't need to convert to int because ceil function already does so automatically

+++ pagination/2-hypermedia_pagination.py
@@ -66,7 +66,7 @@ class Server:
         '''
         data = self.get_page(page, page_size)
         total_items = len(self.dataset())
-        total_pages = int(math.ceil(total_items / page_size))
+        total_pages = math.ceil(total_items / page_size)
         has_prev = page > 1
         has_next = page < total_pages
